+++ s/index.ts
@@ -1,14 +1,14 @@
 
+export * from "./core/circular.js"
 export * from "./core/quat.js"
+export * from "./core/scalar.js"
 export * from "./core/vec2.js"
 export * from "./core/vec3.js"
 export * from "./core/vec4.js"
 
 export * from "./tools/angles.js"
-export * from "./tools/circular.js"
 export * from "./tools/noise.js"
 export * from "./tools/randy.js"
-export * from "./tools/scalar.js"
 export * as spline from "./tools/spline.js"
 
 export * from "./optimizers/hash-map.js"