+++ s/tools/spline.ts
@@ -1,5 +1,5 @@
 
-import {Scalar} from "./scalar.js"
+import {Scalar} from "../core/scalar.js"
 import {XyArray} from "../core/vec2.js"
 
 /** resolve a number within a linear spline. */
@@ -40,50 +40,44 @@ export function catmullRom(x: number, points: XyArray[]) {
 
 		if (x >= x1 && x <= x2) {
 			const t = (x - x1) / (x2 - x1)
-			return helpers.catmullRom(t, points[i - 1], points[i], points[i + 1], points[i + 2])
+			return catmullRomMechanics(t, points[i - 1], points[i], points[i + 1], points[i + 2])
 		}
 	}
 
 	throw new Error("x is out of bounds, try again")
 }
 
-export const ez = {
+/** simple linear spline where the control points are equally-spaced based on their array indices (x is expected to be between 0 and 1). */
+export function ezLinear(x: number, points: number[]) {
+	if (points.length < 2)
+		throw new Error("need at least two points, come on")
 
-	/** simple linear spline where the control points are equally-spaced based on their array indices (x is expected to be between 0 and 1). */
-	linear(x: number, points: number[]) {
-		if (points.length < 2)
-			throw new Error("need at least two points, come on")
+	const points2 = points.map(
+		(p, index): XyArray =>
+			[Scalar.clamp(index / (points.length - 1)), p]
+	)
 
-		const points2 = points.map(
-			(p, index): XyArray =>
-				[Scalar.clamp(index / (points.length - 1)), p]
-		)
-
-		return linear(Scalar.clamp(x), points2)
-	}
+	return linear(Scalar.clamp(x), points2)
 }
 
-namespace helpers {
+/** internal big-brain maths for the catmull-rom implementation */
+function catmullRomMechanics(
+		t: number,
+		[,p0]: XyArray,
+		[,p1]: XyArray,
+		[,p2]: XyArray,
+		[,p3]: XyArray,
+	) {
 
-	/** internal big-brain maths for the catmull-rom implementation */
-	export function catmullRom(
-			t: number,
-			[,p0]: XyArray,
-			[,p1]: XyArray,
-			[,p2]: XyArray,
-			[,p3]: XyArray,
-		) {
+	const t2 = t * t
+	const t3 = t2 * t
 
-		const t2 = t * t
-		const t3 = t2 * t
+	// coefficients for the cubic polynomial (Catmull-Rom)
+	const a = -0.5 * p0 + 1.5 * p1 - 1.5 * p2 + 0.5 * p3
+	const b = p0 - 2.5 * p1 + 2 * p2 - 0.5 * p3
+	const c = -0.5 * p0 + 0.5 * p2
+	const d = p1
 
-		// coefficients for the cubic polynomial (Catmull-Rom)
-		const a = -0.5 * p0 + 1.5 * p1 - 1.5 * p2 + 0.5 * p3
-		const b = p0 - 2.5 * p1 + 2 * p2 - 0.5 * p3
-		const c = -0.5 * p0 + 0.5 * p2
-		const d = p1
-
-		return a * t3 + b * t2 + c * t + d
-	}
+	return a * t3 + b * t2 + c * t + d
 }